services/async_worker/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
autopep8==2.0.2
    # via -r requirements.in
bcrypt==4.0.1
    # via passlib
blis==0.7.9
    # via thinc
boto3==1.26.121
    # via -r requirements.in
botocore==1.29.121
    # via
    #   boto3
    #   s3transfer
catalogue==2.0.8
    # via
    #   spacy
    #   srsly
    #   thinc
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
    #   uvicorn
    #   wasabi
confection==0.0.4
    # via thinc
cryptography==40.0.2
    # via python-jose
cymem==2.0.7
    # via
    #   preshed
    #   spacy
    #   thinc
dataclasses==0.6
    # via -r requirements.in
dnspython==2.3.0
    # via
    #   email-validator
    #   pymongo
ecdsa==0.18.0
    # via python-jose
email-validator==2.0.0.post2
    # via pydantic
exceptiongroup==1.1.1
    # via pytest
fastapi==0.95.1
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.17.0
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via spacy
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
langcodes==3.3.0
    # via spacy
langdetect==1.0.7
    # via spacy-langdetect
markupsafe==2.1.2
    # via jinja2
motor==3.1.2
    # via -r requirements.in
murmurhash==1.0.9
    # via
    #   preshed
    #   spacy
    #   thinc
numpy==1.24.3
    # via
    #   blis
    #   spacy
    #   thinc
packaging==23.1
    # via
    #   pytest
    #   spacy
    #   thinc
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pathy==0.10.1
    # via spacy
pillow==9.5.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
preshed==3.0.8
    # via
    #   spacy
    #   thinc
psycopg2-binary==2.9.6
    # via -r requirements.in
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycodestyle==2.10.0
    # via autopep8
pycparser==2.21
    # via cffi
pydantic[email]==1.10.7
    # via
    #   -r requirements.in
    #   confection
    #   fastapi
    #   spacy
    #   thinc
pymongo==4.3.3
    # via
    #   -r requirements.in
    #   motor
pytest==7.3.1
    # via spacy-langdetect
python-dateutil==2.8.2
    # via botocore
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via -r requirements.in
pyyaml==6.0
    # via uvicorn
requests==2.29.0
    # via
    #   -r requirements.in
    #   spacy
rsa==4.9
    # via python-jose
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via
    #   ecdsa
    #   langdetect
    #   python-dateutil
smart-open==6.3.0
    # via
    #   pathy
    #   spacy
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
spacy==3.5.2
    # via -r requirements.in
spacy-langdetect==0.1.2
    # via -r requirements.in
spacy-legacy==3.0.12
    # via spacy
spacy-loggers==1.0.4
    # via spacy
sqlalchemy==2.0.11
    # via -r requirements.in
srsly==2.4.6
    # via
    #   confection
    #   spacy
    #   thinc
starlette==0.26.1
    # via fastapi
thinc==8.1.9
    # via spacy
tomli==2.0.1
    # via
    #   autopep8
    #   pytest
tqdm==4.65.0
    # via spacy
typer==0.7.0
    # via
    #   pathy
    #   spacy
typing-extensions==4.5.0
    # via
    #   pydantic
    #   sqlalchemy
    #   starlette
urllib3==1.26.15
    # via
    #   botocore
    #   requests
uvicorn[standard]==0.21.1
    # via -r requirements.in
wasabi==1.1.1
    # via
    #   spacy
    #   thinc
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.2
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
